+++ base/sort.jl
@@ -81,7 +81,7 @@ function select!(v::AbstractVector, k::Int, lo::Int, hi::Int, o::Ordering)
     return v[lo]
 end
 
-function select!(v::AbstractVector, r::Range1, lo::Int, hi::Int, o::Ordering)
+function select!(v::AbstractVector, r::UnitRange, lo::Int, hi::Int, o::Ordering)
     a, b = first(r), last(r)
     lo <= a <= b <= hi || error("select index $k is out of range $lo:$hi")
     @inbounds while true
@@ -111,12 +111,12 @@ function select!(v::AbstractVector, r::Range1, lo::Int, hi::Int, o::Ordering)
     end
 end
 
-select!(v::AbstractVector, k::Union(Int,Range1), o::Ordering) = select!(v,k,1,length(v),o)
-select!(v::AbstractVector, k::Union(Int,Range1);
+select!(v::AbstractVector, k::Union(Int,UnitRange), o::Ordering) = select!(v,k,1,length(v),o)
+select!(v::AbstractVector, k::Union(Int,UnitRange);
     lt::Function=isless, by::Function=identity, rev::Bool=false, order::Ordering=Forward) =
     select!(v, k, ord(lt,by,rev,order))
 
-select(v::AbstractVector, k::Union(Int,Range1); kws...) = select!(copy(v), k; kws...)
+select(v::AbstractVector, k::Union(Int,UnitRange); kws...) = select!(copy(v), k; kws...)
 
 # reference on sorted binary search:
 #   http://www.tbray.org/ongoing/When/200x/2003/03/22/Binary